fix: don't use lodash ternary, change subnetID type
can't use the lodash ternary for booleans because it is a pointer. if we
pass the value of the pointer to the ternary function because the value
may not exist. so removed it entirely to simplify how input variables
are defaulted.

change the subnetID type to a pulumi input type so that the output can
be passed from the vpc module.