feat: initial eks component resource, upgrade mods, boilerplate fixes (#3)
* fix: upgrade aws sdk to v5, fix references to boilerplate

* feat: initial eks component resource

left some TODOs that should be addressed, committing now to save
progress and begin some testing.

upgraded all of the modules because generation due to some version
mismatching between the pulumi sdk and pulumi pkg

* chore: document fields

* fix: require nodegroup configuration

* actions: don't build provider if sdk changes

* chore: apply automatic generated SDK changes

* fix: don't use lodash ternary, change subnetID type

can't use the lodash ternary for booleans because it is a pointer. if we
pass the value of the pointer to the ternary function because the value
may not exist. so removed it entirely to simplify how input variables
are defaulted.

change the subnetID type to a pulumi input type so that the output can
be passed from the vpc module.

* fix: add pulumi parent component to all eks resources

* fix: go mod tidy

* fix: missing resource parent, ecr access policy name

* fix: require version 5 of the aws sdk

for passing the eks and oidc object back to the client in the eks module

* chore: apply automatic generated SDK changes

* fix: remove dump-context step, use automation_pat

Co-authored-by: benjyiw <benjyiw@users.noreply.github.com>